src/api/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bibtexparser==1.4.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-cors-headers==4.0.0
django-filebrowser-no-grappelli==4.0.2
django-filter==23.3
django-nested-admin==4.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
django-storages==1.13.2
django-tinymce==3.6.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-writable-nested==0.7.0
drf-yasg==1.21.7
feedparser==6.0.10
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.3
packaging==23.2
Pillow==9.5.0
protobuf==4.23.2
pyparsing==3.0.9
pysolr==3.9.0
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzotero==1.5.9
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.4
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
